% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean Algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{A numeric scalar.}

\item{b}{A numeric scalar.}
}
\value{
The greatest common divisor of a and b.
}
\description{
Computes the greatest common divisor of two numbers
using the Euclidean algorithm.
}
\references{
https://en.wikipedia.org/wiki/Euclidean_algorithm
}
