% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example Graph for Dijkstra's Algorithm}
\format{
A data frame with 18 rows and 3 variables:
\describe{
  \item{v1}{The starting node of an edge.}
  \item{v2}{The destination node of an edge.}
  \item{w}{The weight of the edge.}
}
}
\source{
https://en.wikipedia.org/wiki/Dijkstra%27s_algorithm
}
\usage{
wiki_graph
}
\description{
A weighted graph used to demonstrate Dijkstra's
shortest path algorithm.
}
\keyword{datasets}
