% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/euclidean.R
\name{euclidean}
\alias{euclidean}
\title{Euclidean algorithm}
\usage{
euclidean(a, b)
}
\arguments{
\item{a}{An integer up to 2^53.}

\item{b}{An integer up to 2^53.}
}
\value{
An integer representing the greatest common divisor of the two input
numbers.
}
\description{
Find the greatest common divisor of two numbers.
See \url{https://en.wikipedia.org/wiki/Euclidean_algorithm} for details.
}
\examples{
euclidean(123612, 13892347912)
euclidean(100, 1000)
}
