% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{wiki_graph}
\alias{wiki_graph}
\title{Example graph for Dijkstra algorithm}
\format{
## `wiki_graph`
A graph with 6 vertices:
\describe{
  \item{v1, v2}{Edges given as a pair of vertices}
  \item{w}{Weights}
  ...
}
}
\usage{
wiki_graph
}
\description{
Small example graph
}
\references{
\url{https://en.wikipedia.org/wiki/Dijkstra\%27s_algorithm}
}
\keyword{datasets}
